% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\docType{data}
\name{DVN}
\alias{DVN}
\title{GOF with Facet}
\format{
An object of class \code{character} of length 1.
}
\usage{
DVN
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF with stratification
}
\examples{
lh_dv_strat(...)
}
\keyword{lh_dv_strat}
