% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{forest.dat2}
\alias{forest.dat2}
\title{Prepare dataset for Forest plot data for Forestplot package}
\usage{
forest.dat2(
  data = data,
  param = c("auclast", "cmax", "clast.obs"),
  label,
  covar = c("hs", "sexn"),
  categ = list(c("1=HV", "2=CMV"), c("0=Male", "1=Female"))
)
}
\arguments{
\item{data}{data frame (prepare categorical covariates before hand)}

\item{param}{list of parameters}

\item{label}{label rename}

\item{covar}{define statistics for mid, lower and upper.}

\item{categ}{redefine and order categories. Note that the first category will be used as reference}
}
\description{
Generate dataset for coveffectsplot
}
\keyword{forest.dat2}
