% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{gpairs}
\alias{gpairs}
\title{CONTINUOUS GP EXTERNAL FUNCTION FOR INTERNAL USE}
\usage{
gpairs(
  x,
  upper.pars = list(scatter = "points", conditional = "barcode", mosaic = "mosaic"),
  lower.pars = list(scatter = "points", conditional = "boxplot", mosaic = "mosaic"),
  diagonal = "default",
  outer.margins = list(bottom = unit(2, "lines"), left = unit(2, "lines"), top = unit(2,
    "lines"), right = unit(2, "lines")),
  xylim = NULL,
  outer.labels = NULL,
  outer.rot = c(0, 90),
  gap = 0.05,
  buffer = 0.02,
  reorder = NULL,
  cluster.pars = NULL,
  stat.pars = NULL,
  scatter.pars = NULL,
  bwplot.pars = NULL,
  stripplot.pars = NULL,
  barcode.pars = NULL,
  mosaic.pars = NULL,
  axis.pars = NULL,
  diag.pars = NULL,
  whatis = FALSE
)
}
\arguments{
\item{x}{not to be used internally. Could be downloaded from Github}
}
\description{
Generate COVAR
}
\keyword{gpair}
