% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_boxplot2}
\alias{lh_boxplot2}
\title{BOXPLOT MC SIMULATION}
\usage{
lh_boxplot2(
  data,
  y,
  x,
  x.title,
  y.title,
  target = c(1, 5),
  add.obs.point = "no",
  add.stats = "no",
  stat.label.space = c(0.1, 0.2),
  jit = c(0.1, 0.1),
  targ.line.size = 1,
  targ.text.size = 3,
  targ.text.col = c("blue", "green4", "red"),
  targ.line.col = c("blue", "red"),
  color.target.cuttoff = 80,
  outlier = "no",
  transparency = 0.5,
  obs.color = "grey",
  obs.size = 1,
  target.line.type = c("solid", "dashed"),
  legend.title = "Target Attainment",
  x.label.angle = 45
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate boxplot with targets and stats
}
\examples{
lh_boxplot2(data=rall,
y="Cmaxss",
x="Label",
x.title="Group",
y.title="AUCss",
low.targ.line=rf$Cmaxsslow,
high.targ.line=rf$Cmaxsshi,
add.target="yes",
add.obs.point="yes", tips: GE>= (\+U2265), LE<= (\+U2264)
add.stats="yes",
stat.label.space=c(0.1,0.2),jit=c(0.15,0)) +theme_bw()+ theme(axis.text.x #' @examples = element_text(angle = 45, hjust = 1))+scale_y_log10()
}
\keyword{lh_boxplot2}
