% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cat_box}
\alias{lh_cat_box}
\title{General Boxplots}
\usage{
lh_cat_box(
  data,
  lst.eta = c("ETA1"),
  lst.cov = c("SEX", "RACE"),
  save.path = NULL,
  fancy = "yes",
  hline = F,
  laby = "Individual Random Effect",
  labx = "",
  txt.angle = 45
)
}
\arguments{
\item{data}{Data frame, merged ETA and COVAR data}

\item{lst.eta}{List of ETA names}

\item{lst.cov}{List of covariate names. Plots are generated in loop if more than one covariate}
}
\description{
General Boxplots
}
\examples{
p1<-lh_cat_box(data=cateta,lst.eta=keta,lst.cov=cat,save.path=NULL)
p2<-lh_gof()
}
\keyword{lh_cat_box}
