% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cat_cov}
\alias{lh_cat_cov}
\title{CATEGORICAL BOXPLOT vs ETA}
\usage{
lh_cat_cov(
  data = cateta,
  lst.eta = keta,
  lst.cov = cat,
  save.path = NULL,
  fancy = "yes"
)
}
\arguments{
\item{data}{Data frame, merged ETA and COVAR data}

\item{lst.eta}{List of ETA names}

\item{lst.cov}{List of covariate names. Plots are generated in loop if more than one covariate}
}
\description{
Generate COVAR
}
\examples{
p1<-lh_cat_cov(data=cateta,lst.eta=keta,lst.cov=cat,save.path=NULL)
p2<-lh_gof()
}
\keyword{lh_cat_cov}
