% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_catcon_cov}
\alias{lh_catcon_cov}
\title{CATEGORICAL vs CONTINUOUS BOXPLOT}
\usage{
lh_catcon_cov(
  data,
  lst.eta = c("contcov"),
  con.cov.group = "demo",
  lst.cov = c("SEX", "RACE"),
  save.path = NULL,
  fancy = NULL,
  ytit = "Continuous Covariate",
  xtit = ""
)
}
\arguments{
\item{data}{Data frame, merged ETA and COVAR data}

\item{lst.eta}{List of continuous cov}

\item{lst.cov}{Cat cov names.Plots are generated in loop if more than one covariate.}

\item{save.path}{required if more than 1 Cat cov.}
}
\description{
Generate COVAR CAT vs CONT
}
\examples{
p1<-lh_cat_cov(data=cateta,lst.eta=keta,lst.cov=cat,save.path=NULL)
p2<-lh_gof()
}
\keyword{lh_catcon_cov}
