% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_concat_lattice}
\alias{lh_concat_lattice}
\title{Continuous vs Categorical}
\usage{
lh_concat_lattice(
  dat,
  cont = c("WT", "CRCL", "BSA", "BMI", "AGE"),
  cat = c("CRCLCAT", "NCI", "ECOG", "TUMTYPE"),
  labcont = c("Weight (kg)", "CRCL (mL/min)", "BSA (m^2)", "BMI (kg/m^2)", "Age (years)"),
  labcat = c("Renal\\n Function", "Hepatic\\n Function (NCI)", "ECOG", "Tumor\\n type"),
  xtit = "",
  ytit = "",
  thm = theme(axis.text.x = element_text(vjust = 0.8, hjust = 0.1, size = 9, angle = 90))
)
}
\description{
Plot continuous vs categorical covariates
}
\keyword{lh_concat_lattice}
