% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cwres_dist}
\alias{lh_cwres_dist}
\title{DISTRIBUTION OF CWRES}
\usage{
lh_cwres_dist(data, cwres = "cwres", file.name = "dist_QQ_CWRES.png")
}
\arguments{
\item{data}{Data frame, merged ETA and COVAR data}
}
\description{
Generate GOF
}
\examples{
p1<-lh_con_cov(data=cateta,lst.eta=keta,lst.cov=cat,save.path=NULL)
}
\keyword{lh_cwres_dist}
