% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cwres_pred}
\alias{lh_cwres_pred}
\title{CWRES vs PRED}
\usage{
lh_cwres_pred(
  data = dat1,
  y = "CWRES",
  x = "TIME",
  type = "log",
  scale = c(0.1, 100),
  IPREDN = "Individual Predicted Concentration (ng/mL)",
  PREDN = "Population Predicted Concentration (ng/mL)",
  DVN = "Observed Concentration (ng/mL)",
  TADN = "Time After Dose (h)",
  RTIMEN = "Time After First Dose (h)",
  IVARN = "Time After First Dose (h)",
  CWRESN = "Conditional Weighted Residuals"
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
p1<-lh_cwres_pred(dat1)
p2<-lh_cwres_pred(dat1)
}
\keyword{lh_cwres_pred}
