% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cwres_strat}
\alias{lh_cwres_strat}
\title{CWRES with Facet}
\usage{
lh_cwres_strat(
  data,
  x,
  y,
  id,
  xlab = "Population Predictions (μg/mL)",
  ylab = "Conditional Weighted Residual",
  by = "Age_Group"
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF with stratification
}
\examples{
lh_cwres_strat(...)
}
\keyword{lh_cwres_strat}
