% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cwres_tad}
\alias{lh_cwres_tad}
\title{CWRES vs TAD}
\usage{
lh_cwres_tad(
  data,
  y = "CWRES",
  x = "TAD",
  type = "log",
  scale = c(0.1, 100),
  IPREDN = "Individual Predicted Concentration (ng/mL)",
  PREDN = "Population Predicted Concentration (ng/mL)",
  DVN = "Observed Concentration (ng/mL)",
  TADN = "Time After Dose (h)",
  RTIMEN = "Time After First Dose (h)",
  IVARN = "Time After First Dose (h)",
  CWRESN = "Conditional Weighted Residuals",
  col.obs = "#A6CEE3",
  col.ident = "#1F78B4"
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
data$TAD<-data$TAD;p1<-lh_cwres_tad(dat1)
p2<-lh_cwres_tad(dat1)
}
\keyword{lh_cwres_tad}
