% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_cwres_x}
\alias{lh_cwres_x}
\title{CWRES vs X with Strat}
\usage{
lh_cwres_x(
  data,
  y = "CWRES",
  x = "TIME",
  type = "log",
  scale = c(0.1, 100),
  PREDN = "Population Predicted Concentration (ng/mL)",
  CWRESN = "Conditional Weighted Residuals",
  col.obs = "#A6CEE3",
  col.ident = "#1F78B4",
  strat = NULL
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
p1<-lh_cwres_pred(dat1)
p2<-lh_cwres_pred(dat1)
}
\keyword{lh_cwres_pred}
