% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_dv}
\alias{lh_dv}
\title{Plot DV vs predicted concentration}
\usage{
lh_dv(
  data,
  y = "DV",
  x = "IPRED",
  type = "log",
  scale = "auto",
  xtit = "Individual Predicted Concentration (ng/mL)",
  ytit = "Observed Concentration (ng/mL)",
  sortby = NULL,
  col.obs = "#A6CEE3",
  col.ident = "#1F78B4",
  brew.col = NULL
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate plot with sort color of observe
}
\examples{
lh_dv(...)

}
\keyword{lh_dv}
