% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_dv_ipred}
\alias{lh_dv_ipred}
\title{DV vs IPRED}
\usage{
lh_dv_ipred(
  data,
  y = "DV",
  x = "IPRED",
  type = "log",
  scale = c(0.1, 100),
  IPREDN = "Individual Predicted Concentration (ng/mL)",
  PREDN = "Population Predicted Concentration (ng/mL)",
  DVN = "Observed Concentration (ng/mL)",
  TADN = "Time After Dose (h)",
  RTIMEN = "Time After First Dose (h)",
  IVARN = "Time After First Dose (h)",
  CWRESN = "Conditional Weighted Residuals"
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
p1<-lh_dv_ipred(dat1,type="lin",IPREDN="IPRED")
p2<-lh_dv_ipred(dat1,type="log")
}
\keyword{lh_dv_ipred}
