% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_dv_pred}
\alias{lh_dv_pred}
\title{DV vs PRED}
\usage{
lh_dv_pred(
  data = dat1,
  y = "DV",
  x = "PRED",
  type = "log",
  scale = c(0.1, 100),
  IPREDN = "Individual Predicted Concentration (ng/mL)",
  PREDN = "Population Predicted Concentration (ng/mL)",
  DVN = "Observed Concentration (ng/mL)",
  TADN = "Time After Dose (h)",
  RTIMEN = "Time After First Dose (h)",
  IVARN = "Time After First Dose (h)",
  CWRESN = "Conditional Weighted Residuals",
  col.obs = "#A6CEE3",
  col.ident = "#1F78B4"
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
p1<-lh_dv_pred(dat1,type="lin",IPREDN="IPRED")
p2<-lh_dv_pred(dat1,type="log",scale="auto")
}
\keyword{lh_dv_pred}
