% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_dv_strat}
\alias{lh_dv_strat}
\title{GOF with Facet}
\usage{
lh_dv_strat(
  data,
  x,
  y,
  id,
  xlab = "Individual Predictions (<U+03BC>g/mL)",
  ylab = "Observed\\n Concentrations (<U+03BC>g/mL)",
  by = "Age_Group"
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF with stratification
}
\examples{
lh_dv_strat(...)
}
\keyword{lh_dv_strat}
