% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_dv_x}
\alias{lh_dv_x}
\title{DV vs X with STRAT}
\usage{
lh_dv_x(
  data,
  y = "DV",
  x = "IPRED",
  type = "log",
  scale = c(0.1, 100),
  IPREDN = "Individual Predicted Concentration (ng/mL)",
  DVN = "Observed Concentration (ng/mL)",
  col.obs = "#A6CEE3",
  col.ident = "#1F78B4",
  strat = NULL
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
p1<-lh_dv_ipred(dat1,type="lin",IPREDN="IPRED")
p2<-lh_dv_ipred(dat1,type="log")
}
\keyword{lh_dv_ipred}
