% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_explor_ind}
\alias{lh_explor_ind}
\title{EXPLORATORY Individual plots}
\usage{
lh_explor_ind(
  data,
  dose = "amt",
  id = "id",
  n.plots.page = 9,
  time = "time",
  dv = "dv",
  ipred = NULL,
  pred = NULL,
  type = "linear",
  xtit = "Time after first dose (h)",
  ytit = "Concentration (ng/mL)",
  output.name = "./test.docx"
)
}
\arguments{
\item{data}{Data frame}

\item{output.name}{Listing is generated in word document.}
}
\description{
Generate DATA
}
\examples{
lh_explor_ind(data=dat1,id="usubjid",n.plots.page=9,time="time",dv="dv",ipred="ipred"#' #' @examples ,pred="pred",type="linear",
xtit="Time after first dose (h)",
}
\keyword{lh_indiv_plot}
