% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_indiv_plot}
\alias{lh_indiv_plot}
\title{Individual plots}
\usage{
lh_indiv_plot(
  data = res1,
  id = "id",
  n.plots.page = 9,
  time = "IVAR",
  dv = "DV",
  ipred = "IPRED",
  pred = "PRED",
  plot.obs.type = "point",
  scale = "fixed",
  xxis = list(anglex = 0, sizex = 12, vjustx = 0, strip_size = 12, axis.title.size = 14),
  legendxy = list(posit = "bottom"),
  xtit = "Time after first dose (h)",
  ytit = "Concentration (ng/mL)",
  out.path = "./ind.plots",
  break2 = c(1e-04, 5e-04, 0.001, 0.005, 0.1, 0.5, 1, 5, 10, 100, 10^3, 10^4, 10^5, 10^6),
  p.dpi = 300,
  p.width = 6,
  p.height = 6
)
}
\arguments{
\item{data}{Data frame}

\item{output.name}{Listing is generated in word document.}
}
\description{
Generate GOF
}
\examples{
plh_indiv_plot(data=dat1,id="usubjid",n.plots.page=9,time="time",dv="dv",ipred="ipred"#' #' @examples ,pred="pred",type="linear",
xtit="Time after first dose (h)",
}
\keyword{lh_indiv_plot}
