% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lhboxplot2}
\alias{lhboxplot2}
\title{BOXPLOT MC SIMULATION}
\usage{
lhboxplot2(
  data,
  y,
  x,
  x.title,
  y.title,
  low.targ.line,
  high.targ.line,
  add.target = "no",
  add.obs.point = "no",
  add.stats = "no",
  stat.label.space = c(0.1, 0.2),
  jit = c(0.1, 0.1)
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate boxplot with targets and stats
}
\examples{
lhboxplot2(data=rall,
y="Cmaxss",
x="Label",
x.title="Group",
y.title="AUCss",
low.targ.line=rf$Cmaxsslow,
high.targ.line=rf$Cmaxsshi,
add.target="yes",
add.obs.point="yes",
add.stats="yes",
stat.label.space=c(0.1,0.2),jit=c(0.15,0)) +theme_bw()+ theme(axis.text.x #' @examples = element_text(angle = 45, hjust = 1))+scale_y_log10()
}
\keyword{lhboxplot2}
