% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lhloess}
\alias{lhloess}
\title{Estimate LOESS Fit}
\usage{
lhloess(data = base, x = "basemed", y = "CHG", span = 1)
}
\arguments{
\item{data}{data frame}

\item{x}{x values}

\item{y}{y values}
}
\description{
Generate GOF1
}
\examples{
data$TIME<-data$RTIME;p1<-lh_cwres_time(dat1)
}
\keyword{lhloess}
