% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{qqplot.cwres}
\alias{qqplot.cwres}
\title{DISTRIBUTION OF CWRES internal}
\usage{
qqplot.cwres(dat, ...)
}
\arguments{
\item{data}{Data frame, merged ETA and COVAR data}
}
\description{
Generate GOF
}
\examples{
p1<-lh_con_cov(data=cateta,lst.eta=keta,lst.cov=cat,save.path=NULL)
}
\keyword{qqplot.cwres}
