% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{vpc_cat}
\alias{vpc_cat}
\title{VPC with Categorical X}
\usage{
vpc_cat(
  sim.data = vpc$stats,
  obs.data = vpc$obs,
  obs.var = c("x", "y"),
  box.width = 0.1,
  bin.cat = "xbin",
  y.sim = c("y", "lo", "md", "hi", "qname"),
  xtit = "",
  ytit = "Concentration (ng/mL)",
  log.y = NULL,
  x.lab = c("Pre-dose", "1-h Post-dose")
)
}
\arguments{
\item{sim.data}{Sim dataset resulted from vpc function in tidyvpc}

\item{obs.data}{Obs dataset resulted from vpc function in tidyvp.Set it to null to omit the observed data}

\item{box.width}{Width parameter for PI shaded area}

\item{x.lab}{X axis label in characters, mandatory}
}
\description{
Generate categorical VPC plot
}
\examples{
vpc <- observed(obs, x=NTIMEN, y=DV) \%>\%
simulated (sim, y=DV)\%>\%
tidyvpc::stratify(~DOSE) \%>\%
binning(bin =NTIMEN)\%>\%
p<-vpcstats() + other ggplot functions such as facet_wrap/grid, etc.
vpc_cat(...)
}
