% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_boxplot2}
\alias{lh_boxplot2}
\title{BOXPLOT MC SIMULATION}
\usage{
lh_boxplot2(
  data = NULL,
  y.dat = NULL,
  x.dat = NULL,
  x.title = NULL,
  y.title,
  target = c(1, 5),
  add.obs.point = "no",
  add.stats = "no",
  stat.label.space = c(0.1, 0.2),
  max.y = NULL,
  min.y = 0,
  increm.y = 100,
  box.stats = c("quantile(x,0.05)", "quantile(x,0.25)", "quantile(x,0.5)",
    "quantile(x,0.75)", "quantile(x,0.95)"),
  jit = c(0.1, 0.1),
  targ.line.size = 1,
  targ.text.size = 3,
  targ.text.col = c("blue", "green4", "red"),
  targ.line.col = c("blue", "red"),
  color.target.cuttoff = 80,
  outlier = "no",
  transparency = 0.5,
  obs.color = "grey",
  obs.size = 1,
  target.line.type = c("solid", "dashed"),
  legend.title = "Target Attainment",
  targ.legend = c("missed", "below adult geomeans at 400 mg BID",
    "between adult geomeans at 400 & 1200 mg BID",
    "Simulated median above adult geomean at 1200 mg BID")
)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate boxplot with targets and stats
}
\examples{
lh_boxplot2(data=rall,
y.dat="Cmaxss",
x.dat="Label",
x.title="Group",
y.title="AUCss",
low.targ.line=rf$Cmaxsslow,
high.targ.line=rf$Cmaxsshi,
add.target="yes",
add.obs.point="yes", tips: GE>= (\+U2265), LE<= (\+U2264)
add.stats="yes",
stat.label.space=c(0.1,0.2),jit=c(0.15,0)) +theme_bw()+ theme(axis.text.x #' @examples = element_text(angle = 45, hjust = 1))+scale_y_log10()
}
\keyword{lh_boxplot2}
