% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_con_cov}
\alias{lh_con_cov}
\title{CONTINUOUS SCATTER vs ETA}
\usage{
lh_con_cov(
  data,
  lst.eta = c("eta1", "eta2"),
  lst.cov = c("AGE", "WT"),
  save.path = "./scatter.png",
  fancy = "yes",
  diag.fs = 12,
  stat.fz = 12,
  p.width = 768,
  p.height = 768,
  p.pointsize = 16,
  scat.pch = 20
)
}
\arguments{
\item{data}{Data frame, merged ETA and COVAR data}

\item{lst.eta}{List of ETA names}

\item{lst.cov}{List of covariate names. Add plot name to save.path}
}
\description{
Generate COVAR
}
\examples{
p1<-lh_con_cov(data=cateta,lst.eta=keta,lst.cov=cat,save.path=NULL)
}
\keyword{lh_con_cov}
