% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_gof}
\alias{lh_gof}
\title{SAVE BATCH PLOTS}
\usage{
lh_gof(file.name = NULL)
}
\arguments{
\item{data}{data frame}
}
\description{
Generate GOF1
}
\examples{
p1<-lh_gof("test.png")
p2<-lh_gof()
}
\keyword{lh_gof}
