% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhplot.R
\name{lh_indiv_plot}
\alias{lh_indiv_plot}
\title{Individual plots}
\usage{
lh_indiv_plot(
  data = dat1,
  id = "usubjid",
  n.plots.page = 9,
  time = "time",
  dv = "dv",
  ipred = "ipred",
  pred = "pred",
  type = "linear",
  xtit = "Time after first dose (h)",
  ytit = "Concentration (ng/mL)",
  output.name = "./test.docx"
)
}
\arguments{
\item{data}{Data frame}

\item{output.name}{Listing is generated in word document.}
}
\description{
Generate GOF
}
\examples{
plh_indiv_plot(data=dat1,id="usubjid",n.plots.page=9,time="time",dv="dv",ipred="ipred"#' #' @examples ,pred="pred",type="linear",
xtit="Time after first dose (h)",
}
\keyword{lh_indiv_plot}
